Dohyun

+++ src/main/java/service/ItemService.java
@@ -1,5 +1,6 @@
 package service;
 
+import common.UserInput;
 import common.Role;
 import common.UserInput;
 import domain.Item;
@@ -10,6 +11,8 @@ import repository.*;
 
 import java.util.Scanner;
 import java.util.stream.Collectors;
+
+import static common.UserInput.*;
 
 import static common.UserInput.*;
 
@@ -35,6 +38,11 @@ public class ItemService {
         displayItemMenu();
         int choice = inputInt("선택: ", sc);
 
+        switch (choice){
+            case 1: //상품 관련 메인 서비스 호출 (카테고리별 상품 조회, 장바구니, 특정 상품 리뷰)
+                //(주방 > 냄비 ,후라이 등등. .> 제품 선택하면 > 제품 데이터 + (제품 id에 대한)리뷰, 장바구니 담기)
+                handleItemMainService(sc);
+
         switch (choice) {
             case 1: //카테고리별 조회 서비스 호출
                 int middleNum = categoryItemViewService();
@@ -48,117 +56,6 @@ public class ItemService {
                 //리뷰 (1. 리뷰 서비스) => reviewService(itemId)
                 break;
         }
-
-//        switch (choice) {
-//            case 1 :
-//                majorCategory();
-//                int categoryChoice = sc.nextInt();
-//                switch (categoryChoice) {
-//                    int subChoice;
-//                    String keyword;
-//                    case 1 :
-//                        subChoice = sc.nextInt();
-//                        electronMiddleCategory();
-//                        switch (subChoice){
-//                            case 1 :
-//                                keyword = "냉장고";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 2 :
-//                                keyword = "TV";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 3 :
-//                                keyword = "세탁기";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 4 :
-//                                keyword = "청소기";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 5 :
-//                                keyword = "전자레인지";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 6 :
-//                                keyword = "컴퓨터";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                        }
-//                        break;
-//                    case 2 :
-//                        petMiddleCategory();
-//                        subChoice = sc.nextInt();
-//                        switch (subChoice) {
-//                            case 1:
-//                                keyword = "사료";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 2:
-//                                keyword = "간식";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 3:
-//                                keyword = "용품";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                        }
-//                        break;
-//                    case 3 :
-//                        clothesMiddleCategory();
-//                        subChoice = sc.nextInt();
-//                        switch (subChoice) {
-//                            case 1:
-//                                keyword = "상의";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 2:
-//                                keyword = "하의";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 3:
-//                                keyword = "신발";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 4:
-//                                keyword = "속옷";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 5:
-//                                keyword = "점퍼";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                            case 6:
-//                                keyword = "가방/악세서리";
-//                                itemRepository.findByKeyword(keyword);
-//                                break;
-//                        }
-//                    case 4 :
-//                        foodMiddleCategory();
-//                        break;
-//                    case 5 :
-//                        livingMiddleCategory();
-//                        break;
-//                    case 6 :
-//                        sportsMiddleCategory();
-//                        break;
-//                    case 7 :
-//                        phraseMiddleCategory();
-//                        break;
-//                    case 8 :
-//                        bookMiddleCategory();
-//                        break;
-//                }
-//                break;
-//            case 2 :
-//                System.out.println("검색어를 입력해주세요.");
-//                String keyword = sc.next();
-//                itemRepository.findByKeyword(keyword);
-//                break;
-//            case 3 :
-//                handleReviewService(sc);
-//                break;
-//        }
     }
 
     private void itemSearchService() {
@@ -174,90 +71,6 @@ public class ItemService {
 
     private static boolean isAdmin(Member currentMember) {
         return currentMember.getRole().equals(Role.ADMIN);
-    }
-
-    private static void bookMiddleCategory() {
-        System.out.println("===== 중분류 =====");
-        System.out.println("1. 소설");
-        System.out.println("2. 경제/경영");
-        System.out.println("3. IT");
-        System.out.println("4. 예술");
-        System.out.println("5. 기술/공학");
-        System.out.println("6. 유아");
-    }
-
-    private static void phraseMiddleCategory() {
-        System.out.println("===== 중분류 =====");
-        System.out.println("1. 필기구");
-        System.out.println("2. 노트/메모지");
-        System.out.println("3. 유아");
-    }
-
-    private static void sportsMiddleCategory() {
-        System.out.println("===== 중분류 =====");
-        System.out.println("1. 구기");
-        System.out.println("2. 헬스/요가");
-        System.out.println("3. 라켓스포츠");
-        System.out.println("4. 수영");
-        System.out.println("5. 낚시");
-    }
-
-    private static void livingMiddleCategory() {
-        System.out.println("===== 중분류 =====");
-        System.out.println("1. 잡화");
-        System.out.println("2. 세제");
-        System.out.println("3. 수납/정리");
-        System.out.println("4. 바디/헤어/구강/면도");
-        System.out.println("5. 가구/조명/인테리어");
-    }
-
-    private static void foodMiddleCategory() {
-        System.out.println("===== 중분류 =====");
-        System.out.println("1. 과일");
-        System.out.println("2. 축산");
-        System.out.println("3. 수산물/건어물");
-        System.out.println("4. 냉장/냉동/간편요리");
-        System.out.println("5. 쌀/잡곡");
-        System.out.println("6. 커피/원두/차");
-    }
-
-    private static void clothesMiddleCategory() {
-        System.out.println("===== 중분류 =====");
-        System.out.println("1. 상의");
-        System.out.println("2. 하의");
-        System.out.println("3. 신발");
-        System.out.println("4. 속옷");
-        System.out.println("5. 점퍼");
-        System.out.println("6. 가방/악세서리");
-    }
-
-    private static void petMiddleCategory() {
-        System.out.println("===== 중분류 =====");
-        System.out.println("1. 사료");
-        System.out.println("2. 간식");
-        System.out.println("3. 용품");
-    }
-
-    private static void electronMiddleCategory() {
-        System.out.println("===== 중분류 =====");
-        System.out.println("1. 냉장고");
-        System.out.println("2. TV");
-        System.out.println("3. 세탁기");
-        System.out.println("4. 청소기");
-        System.out.println("5. 전자레인지");
-        System.out.println("6. 컴퓨터");
-    }
-
-    private static void majorCategory() {
-        System.out.println("===== 대분류 =====");
-        System.out.println("1. 가전/디지털");
-        System.out.println("2. 반려동물용품");
-        System.out.println("3. 의류");
-        System.out.println("4. 식품");
-        System.out.println("5. 생활용품");
-        System.out.println("6. 스포츠/레저");
-        System.out.println("7. 문구/완구");
-        System.out.println("8. 도서");
     }
 
     private static void displayUserMenu() {
@@ -288,7 +101,7 @@ public class ItemService {
 
         switch (choice){
             case 1: //리뷰보기
-                selectReview();
+                selectReview(sc);
                 break;
             case 2: //리뷰작성
                 insertReview(sc);
@@ -304,40 +117,75 @@ public class ItemService {
         }
     }
 
+    // (제품 아이디별) 
+    private void selectReview(Scanner sc) {
+        Long itemId = inputLong("리뷰 보실 제품의 아이디를 입력해 주세요 :", sc);
+        System.out.println("""
+                정렬기준을 선택해 주세요.
+                1. 최신순
+                2. 별점 높은순
+                3. 별점 낮은순
+                """);
+        int sortOption = sc.nextInt();
+        int sort;
+        switch (sortOption) {
+            case 1:
+                sort = 1;
+                break;
+            case 2:
+                sort = 2;
+                break;
+            case 3:
+                sort = 3;
+                break;
+            default:
+                System.out.println("잘못된 입력입니다. 기본값으로 최신순으로 정렬됩니다.");
+                sort = 1;
+                break;
+        }
+        reviewRepository.findByAll(itemId, sort);
+
     private void selectReview() {
         // 권한 확인
         Long memberId = Session.getInstance().getCurrentMember().getMemberId();
         reviewRepository.findByMemberId(memberId);
     }
 
+//    (회원(memberID)별 리뷰보기 완료)
+//    private void memberIdRewview(){
+//        Long memberId = Session.getInstance().getCurrentMember().getMemberId();
+//        reviewRepository.findReviewById(memberId);
+//    }
+
+    // (리뷰 삭제하기 완료)
     private void deleteReview(Scanner sc) {
-        System.out.println("삭제할 리뷰의 아이디를 입력해 주세요.");
-        Long reviewId = sc.nextLong();
+        Long reviewId = inputLong("삭제할 리뷰의 아이디를 입력해 주세요 :", sc);
         // 권한 확인
         Long memberId = Session.getInstance().getCurrentMember().getMemberId();
+        boolean findMemberID = reviewRepository.findById(reviewId, memberId);
 
-
-        reviewRepository.deleteById(reviewId);
+        if(findMemberID){
+            reviewRepository.deleteById(reviewId);
+            System.out.println("리뷰가 삭제 되었습니다.");
+        }else{
+            System.out.println("삭제할 수 있는 리뷰가 없습니다.");
+        }
     }
 
+    // (리뷰 작성하기 완료)
     private void insertReview(Scanner sc) {
-        System.out.println("리뷰를 남기실 상품 ID를 입력해주세요.");
-        Long itemId = sc.nextLong();
+        Long itemId = inputLong("리뷰를 남기실 상품 ID를 입력해주세요:", sc);
         //해당 상품을 구매한 이력이 있는지 체크
         //이것은 리뷰테이블과는 관계가 없는 회원의 아이디
         Long memberId = Session.getInstance().getCurrentMember().getMemberId();
-
         boolean isOrdered = orderItemRepository.existsByItemIdAndMemberId(itemId, memberId);
 
-        //성공 시
         if (isOrdered){
             //성공 시
-            System.out.println("평점을 입력해주세요.");
-            int stars = sc.nextInt();
-            System.out.println("내용을 입력해주세요.");
-            String contents = sc.next();
+            int stars = inputInt("평점을 입력해주세요: ", sc);
+            String contents = inputString("내용을 입력해주세요: ", sc);
 
-            Review review = Review.of(stars, contents,memberId,itemId);
+            Review review = Review.of(stars,contents,memberId,itemId);
             reviewRepository.save(review);
         }else {
             //실패 시
@@ -345,24 +193,24 @@ public class ItemService {
         }
     }
 
+    // (리뷰 수정하기 완료)
     private void updateReview(Scanner sc) {
-//        System.out.println("수정할 리뷰 아이디를 입력해주세요.");
-//        Long reviewId = sc.nextLong();
-//        // 권한 확인
-//        Long memberId = Session.getInstance().getCurrentMember().getMemberId();
-//
-//        Review findReview = reviewRepository.findById(reviewId);
-//
-//        //해당 리뷰 작성자가 본인이 맞을 경우
-//        if (Objects.equals(findReview.getMemberId(), memberId)) {
-//            System.out.println("수정할 별점을 입력해 주세요.");
-//            int stars = sc.nextInt();
-//            System.out.println("수정할 리뷰 내용을 입력해주세요.");
-//            String contents = sc.next();
-//
-//            Review review = Review.of(reviewId, stars, contents);
-//            reviewRepository.updateById(reviewId, review);
-//        }
+        Long reviewId = inputLong("수정할 리뷰의 아이디를 입력해주세요 :", sc);
+        Long memberId = Session.getInstance().getCurrentMember().getMemberId();
+        boolean findMemberID = reviewRepository.findById(reviewId, memberId);
+
+        //해당 리뷰 작성자가 본인이 맞을 경우
+        if (findMemberID) {
+            int stars = inputInt("수정할 별점을 입력해 주세요 :", sc);
+            String contents = inputString("수정할 리뷰 내용을 입력해주세요 :", sc);
+
+            Review review = Review.of(stars, contents);
+            reviewRepository.updateById(reviewId, review);
+            System.out.println("리뷰가 수정 되었습니다.");
+        }
+        else {
+            System.out.println("수정할 리뷰가 없습니다.");
+        }
     }
 
     private static void displayItemMenu() {
@@ -371,7 +219,6 @@ public class ItemService {
                 2. 상품 키워드 검색
                 """);
     }
-
     private static void displayReviewMenu() {
         System.out.println("""
                 1. 리뷰보기
@@ -380,7 +227,8 @@ public class ItemService {
                 4. 리뷰삭제
                 """);
     }
-    
+
+
     private static void serviceBreak() {
         System.out.println("잘못된 번호 입니다.");
     }
